+++ vault_contract/src/issuer.rs
@@ -6,17 +6,17 @@ use soroban_sdk::{contracttype, panic_with_error, Address, Env, Map, String};
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Issuer {
     pub public_key: Address,
-    pub is_revoked: bool,
+    pub revoked: bool,
 }
 
-pub fn authorize_issuer(e: &Env, issuer_pk: &Address, did: &String) {
+pub fn authorize_issuer(e: &Env, issuer: &Address, did: &String) {
     let mut issuers: Map<Address, Issuer> = storage::read_issuers(e, did);
 
     issuers.set(
-        issuer_pk.clone(),
+        issuer.clone(),
         Issuer {
-            public_key: issuer_pk.clone(),
-            is_revoked: false,
+            public_key: issuer.clone(),
+            revoked: false,
         },
     );
 
@@ -31,7 +31,7 @@ pub fn revoke_issuer(e: &Env, issuer: &Address, did: &String) {
             issuer.clone(),
             Issuer {
                 public_key: issuer.clone(),
-                is_revoked: true,
+                revoked: true,
             },
         )
     } else {
@@ -46,5 +46,5 @@ pub fn is_registered(issuers: &Map<Address, Issuer>, issuer: &Address) -> bool {
 }
 
 pub fn is_revoked(issuers: &Map<Address, Issuer>, issuer: &Address) -> bool {
-    issuers.get_unchecked(issuer.clone()).is_revoked
+    issuers.get_unchecked(issuer.clone()).revoked
 }